+++ packages/firebase_performance/example/lib/main.dart
@@ -4,6 +4,7 @@
 
 import 'dart:async';
 
+import 'package:http/http.dart';
 import 'package:flutter/material.dart';
 
 import 'package:firebase_performance/firebase_performance.dart';
@@ -15,12 +16,41 @@ class MyApp extends StatefulWidget {
   _MyAppState createState() => new _MyAppState();
 }
 
+class _MetricHttpClient extends BaseClient {
+  _MetricHttpClient(this._inner);
+
+  final Client _inner;
+
+  @override
+  Future<StreamedResponse> send(BaseRequest request) async {
+    final HttpMetric metric = FirebasePerformance.instance
+        .newHttpMetric(request.url.toString(), HttpMethod.Get);
+
+    await metric.start();
+
+    StreamedResponse response;
+    try {
+      response = await _inner.send(request);
+      metric
+        ..responsePayloadSize = response.contentLength
+        ..responseContentType = response.headers['Content-Type']
+        ..requestPayloadSize = request.contentLength
+        ..httpResponseCode = response.statusCode;
+    } finally {
+      await metric.stop();
+    }
+
+    return response;
+  }
+}
+
 class _MyAppState extends State<MyApp> {
   FirebasePerformance _performance = FirebasePerformance.instance;
   bool _isPerformanceCollectionEnabled = false;
   String _performanceCollectionMessage =
       'Unknown status of performance collection.';
   bool _traceHasRan = false;
+  bool _httpMetricHasRan = false;
 
   @override
   void initState() {
@@ -42,6 +72,10 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future<void> _testTrace() async {
+    setState(() {
+      _traceHasRan = false;
+    });
+
     final Trace trace = _performance.newTrace("test");
     trace.incrementCounter("counter1", 16);
     trace.putAttribute("favorite_color", "blue");
@@ -61,32 +95,58 @@ class _MyAppState extends State<MyApp> {
     });
   }
 
+  Future<void> _testHttpMetric() async {
+    setState(() {
+      _httpMetricHasRan = false;
+    });
+
+    final _MetricHttpClient metricHttpClient =
+        new _MetricHttpClient(new Client());
+    final Request request =
+        new Request("SEND", Uri.parse("https://www.google.com"));
+    metricHttpClient.send(request);
+
+    setState(() {
+      _httpMetricHasRan = true;
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
+    final TextStyle textStyle =
+        const TextStyle(color: Colors.lightGreenAccent, fontSize: 25.0);
     return new MaterialApp(
       home: new Scaffold(
         appBar: new AppBar(
           title: const Text('Firebase Performance Example'),
         ),
         body: new Center(
-            child: new Column(
-          children: <Widget>[
-            new Text(_performanceCollectionMessage),
-            new RaisedButton(
-              onPressed: _togglePerformanceCollection,
-              child: const Text('Toggle Data Collection'),
-            ),
-            new RaisedButton(
-              onPressed: _testTrace,
-              child: const Text('Run Trace'),
-            ),
-            new Text(
-              _traceHasRan ? 'Trace Ran!' : '',
-              style: const TextStyle(
-                  color: Colors.lightGreenAccent, fontSize: 25.0),
-            )
-          ],
-        )),
+          child: new Column(
+            children: <Widget>[
+              new Text(_performanceCollectionMessage),
+              new RaisedButton(
+                onPressed: _togglePerformanceCollection,
+                child: const Text('Toggle Data Collection'),
+              ),
+              new RaisedButton(
+                onPressed: _testTrace,
+                child: const Text('Run Trace'),
+              ),
+              new Text(
+                _traceHasRan ? 'Trace Ran!' : '',
+                style: textStyle,
+              ),
+              new RaisedButton(
+                onPressed: _testHttpMetric,
+                child: const Text('Run HttpMetric'),
+              ),
+              new Text(
+                _httpMetricHasRan ? 'HttpMetric Ran!' : '',
+                style: textStyle,
+              ),
+            ],
+          ),
+        ),
       ),
     );
   }