+++ packages/firebase_performance/lib/firebase_performance.dart
@@ -11,4 +11,5 @@ import 'package:flutter/services.dart';
 import 'package:flutter/foundation.dart';
 
 part 'src/firebase_performance.dart';
+part 'src/http_metric.dart';
 part 'src/trace.dart';