+++ packages/firebase_performance/lib/src/trace.dart
@@ -55,8 +55,7 @@ class Trace {
 
   /// Stops this trace.
   ///
-  /// Using ```await``` with this method is only necessary when accurate timing
-  /// is relevant.
+  /// Not necessary to use ```await``` with this method.
   Future<void> stop() {
     assert(!_hasStopped);
     assert(_hasStarted);
@@ -118,8 +117,8 @@ class Trace {
 
   /// Removes an already added [attribute].
   ///
-  /// If the trace has been stopped, this method returns without removing the
-  /// attribute.
+  /// If the trace has been stopped, this method throws an assertion
+  /// error.
   void removeAttribute(String attribute) {
     assert(!_hasStopped);