+++ packages/url_launcher/ios/Classes/UrlLauncherPlugin.m
@@ -37,8 +37,6 @@
 
 - (void)safariViewControllerDidFinish:(SFSafariViewController *)controller {
   [controller dismissViewControllerAnimated:YES completion:nil];
-  _url = nil;
-  _flutterResult = nil;
 }
 
 @end