+++ packages/firebase_performance/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+* Added HttpMetric for monitoring for specific network requests.
+
 ## 0.0.1
 
 * Initial Release